turo_backend/requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0